requirements.txt
requests
python-dotenv
aiohttp
gql
asyncio
pymongo